feat: update EvaluationPage to set document title with available data; refactor title setting logic feat: change work hours fields in WeeklyRecord model to Float for better precision feat: add migration script to alter work hours columns to DOUBLE PRECISION feat: implement delay when opening evaluation in a new window to ensure sessionStorage is updated

+++ components/AddEditWeeklyRecord.tsx
@@ -79,7 +79,7 @@ export default function AddEditWeeklyRecord({
       ...prev,
       [name]: ["startDate", "endDate"].includes(name)
         ? value
-        : Math.max(0, parseInt(value) || 0),
+        : Math.max(0, parseFloat(value) || 0),
     }));
     // Clear error for this field
     if (errors[name]) {
